Extract MultiParentCasperTestUtil
