+++ casper/src/test/scala/coop/rchain/casper/MultiParentCasperTestUtil.scala
@@ -1,0 +1,126 @@
+package coop.rchain.casper
+
+import java.nio.file.Files
+
+import cats.{Applicative, Monad}
+import cats.data.EitherT
+import cats.effect.{Concurrent, Sync}
+import cats.implicits._
+import com.google.protobuf.ByteString
+import coop.rchain.blockstorage.BlockStore
+import coop.rchain.casper.Estimator.Validator
+import coop.rchain.casper.genesis.Genesis
+import coop.rchain.casper.genesis.contracts._
+import coop.rchain.casper.MultiParentCasper.ignoreDoppelgangerCheck
+import coop.rchain.casper.helper.HashSetCasperTestNode.Effect
+import coop.rchain.casper.helper.{BlockDagStorageTestFixture, BlockUtil, HashSetCasperTestNode}
+import coop.rchain.casper.protocol._
+import coop.rchain.casper.util.{BondingUtil, ProtoUtil}
+import coop.rchain.casper.util.ProtoUtil.{signBlock, toJustification}
+import coop.rchain.casper.util.rholang.RuntimeManager
+import coop.rchain.casper.util.rholang.InterpreterUtil.mkTerm
+import coop.rchain.catscontrib.TaskContrib.TaskOps
+import coop.rchain.comm.rp.ProtocolHelper.packet
+import coop.rchain.comm.{transport, CommError, TimeOut}
+import coop.rchain.crypto.{PrivateKey, PublicKey}
+import coop.rchain.crypto.codec.Base16
+import coop.rchain.crypto.hash.{Blake2b256, Keccak256}
+import coop.rchain.crypto.signatures.{Ed25519, Secp256k1}
+import coop.rchain.p2p.EffectsTestInstances.LogicalTime
+import coop.rchain.rholang.interpreter.{accounting, Runtime}
+import coop.rchain.models.{Expr, Par}
+import coop.rchain.shared.StoreType
+import coop.rchain.shared.PathOps.RichPath
+import coop.rchain.catscontrib._
+import coop.rchain.catscontrib.Catscontrib._
+import coop.rchain.catscontrib.eitherT._
+import monix.eval.Task
+import monix.execution.Scheduler
+import monix.execution.Scheduler.Implicits.global
+import org.scalatest.{Assertion, FlatSpec, Inspectors, Matchers}
+import coop.rchain.casper.scalatestcontrib._
+import coop.rchain.casper.util.comm.TestNetwork
+import coop.rchain.catscontrib.ski.kp2
+import coop.rchain.comm.rp.Connect.Connections
+import coop.rchain.metrics
+import coop.rchain.metrics.Metrics
+import coop.rchain.shared.Log
+import org.scalatest
+
+import scala.collection.immutable
+import scala.util.Random
+import scala.concurrent.duration._
+
+object MultiParentCasperTestUtil {
+  def validateBlockStore[R](
+      node: HashSetCasperTestNode[Effect]
+  )(f: BlockStore[Effect] => Effect[R])(implicit metrics: Metrics[Effect], log: Log[Effect]) =
+    for {
+      bs     <- BlockDagStorageTestFixture.createBlockStorage[Effect](node.blockStoreDir)
+      result <- f(bs)
+      _      <- bs.close()
+      _      <- Sync[Effect].delay { node.blockStoreDir.recursivelyDelete() }
+    } yield result
+
+  def blockTuplespaceContents(
+      block: BlockMessage
+  )(implicit casper: MultiParentCasper[Effect]): Effect[String] = {
+    val tsHash = ProtoUtil.postStateHash(block)
+    MultiParentCasper[Effect].storageContents(tsHash)
+  }
+
+  def deployAndQuery(
+      node: HashSetCasperTestNode[Effect],
+      dd: DeployData,
+      query: DeployData
+  ): Effect[(BlockStatus, Seq[Par])] =
+    for {
+      createBlockResult <- node.casperEff.deploy(dd) >> node.casperEff.createBlock
+      Created(block)    = createBlockResult
+      blockStatus       <- node.casperEff.addBlock(block, ignoreDoppelgangerCheck[Effect])
+      queryResult <- node.runtimeManager
+                      .captureResults(ProtoUtil.postStateHash(block), query)
+    } yield (blockStatus, queryResult)
+
+  def createBonds(validators: Seq[Array[Byte]]): Map[Array[Byte], Long] =
+    validators.zipWithIndex.map { case (v, i) => v -> (2L * i.toLong + 1L) }.toMap
+
+  def createGenesis(bonds: Map[Array[Byte], Long]): BlockMessage =
+    buildGenesis(Seq.empty, bonds, 1L, Long.MaxValue, Faucet.noopFaucet, 0L)
+
+  def buildGenesis(
+      wallets: Seq[PreWallet],
+      bonds: Map[Array[Byte], Long],
+      minimumBond: Long,
+      maximumBond: Long,
+      faucetCode: String => String,
+      deployTimestamp: Long
+  ): BlockMessage = {
+    val initial                            = Genesis.withoutContracts(bonds, 1L, deployTimestamp, "rchain")
+    val storageDirectory                   = Files.createTempDirectory(s"hash-set-casper-test-genesis")
+    val storageSize: Long                  = 1024L * 1024
+    implicit val log                       = new Log.NOPLog[Task]
+    implicit val metricsEff: Metrics[Task] = new metrics.Metrics.MetricsNOP[Task]
+
+    val activeRuntime =
+      Runtime
+        .createWithEmptyCost[Task, Task.Par](storageDirectory, storageSize, StoreType.LMDB)
+        .unsafeRunSync
+    val runtimeManager = RuntimeManager.fromRuntime[Task](activeRuntime).unsafeRunSync
+    val emptyStateHash = runtimeManager.emptyStateHash
+    val validators     = bonds.map(bond => ProofOfStakeValidator(bond._1, bond._2)).toSeq
+    val genesis = Genesis
+      .withContracts(
+        initial,
+        ProofOfStakeParams(minimumBond, maximumBond, validators),
+        wallets,
+        faucetCode,
+        emptyStateHash,
+        runtimeManager,
+        deployTimestamp
+      )
+      .unsafeRunSync
+    activeRuntime.close().unsafeRunSync
+    genesis
+  }
+}
