+++ casper/src/test/scala/coop/rchain/casper/RholangBuildTest.scala
@@ -14,8 +14,8 @@ import org.scalatest.{FlatSpec, Matchers}
 class RholangBuildTest extends FlatSpec with Matchers {
 
   val (validatorKeys, validators) = (1 to 4).map(_ => Ed25519.newKeyPair).unzip
-  val bonds                       = HashSetCasperTest.createBonds(validators)
-  val genesis                     = HashSetCasperTest.createGenesis(bonds)
+  val bonds                       = MultiParentCasperTestUtil.createBonds(validators)
+  val genesis                     = MultiParentCasperTestUtil.createGenesis(bonds)
 
   //put a new casper instance at the start of each
   //test since we cannot reset it
@@ -40,7 +40,7 @@ class RholangBuildTest extends FlatSpec with Matchers {
       createBlockResult    <- MultiParentCasper[Effect].deploy(deploy) *> MultiParentCasper[Effect].createBlock
       Created(signedBlock) = createBlockResult
       _                    <- MultiParentCasper[Effect].addBlock(signedBlock, ignoreDoppelgangerCheck[Effect])
-      storage              <- HashSetCasperTest.blockTuplespaceContents(signedBlock)
+      storage              <- MultiParentCasperTestUtil.blockTuplespaceContents(signedBlock)
       _                    = logEff.warns should be(Nil)
       _                    = storage.contains("!([4, 6, 10, 14])") should be(true)
       result               = storage.contains("!(\"The timestamp is 1\")") should be(true)
